Add test with maximal address, and fix code

+++ r600/cf_node.cpp
@@ -214,7 +214,7 @@ cf_alu_node::cf_alu_node(uint64_t bc, uint64_t bc_ext):
 
 cf_alu_node::cf_alu_node(uint16_t opcode,
                          uint16_t flags,
-                         uint16_t addr,
+                         uint32_t addr,
                          uint16_t count,
                          const std::tuple<int, int,int>& kcache0,
                          const std::tuple<int, int,int>& kcache1
